requirements-dev.txt
pytest
pytest-cov
pytest-pep8
pytest-html
pytest-xdist
pytest-flask
pep8
